pyproject.toml
[project]
name = "stock-market-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.50.0",
    "datetime",
    "finnhub-python>=2.4.23",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.12.3",
    "readability-lxml>=0.8.1",
    "lxml>=5.2.2",
    "google-genai>=1.10.0",
    "pydantic>=2.7.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.3.0",
    "mcp[cli]>=1.6.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "langchain-groq>=1.1.2",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
]
